requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
Django==5.0.1
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-colorfield==0.11.0
django-crispy-forms==2.1
django-js-asset==2.2.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
environs==10.3.0
gunicorn==21.2.0
marshmallow==3.20.2
packaging==23.2
pillow==10.2.0
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2023.3.post1
qrcode==7.4.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzlocal==5.2